More pytorch cuda edits

+++ src/test/py/ltprg/game/synthetic_nom_ref/basic_model.py
@@ -64,9 +64,9 @@ class BasicModel(object):
 
         # dtype
         if cuda.is_available():
-            self.dtype = torch.cuda.LongTensor
+            self.dtype = torch.cuda.FloatTensor
         else:
-            self.dtype = torch.LongTensor
+            self.dtype = torch.FloatTensor
 
 
     def evaluate(self, prediction, label):
